backend/requirements.txt
fastapi
uvicorn
markdown
python-frontmatter
pydantic[email]
openai
python-dotenv
python-multipart
google-genai
sqlalchemy
psycopg2-binary
alembic
gunicorn
google-cloud-storage
python-jose[cryptography]
pyjwt
requests